+++ lib/ex_plasma/block.ex
@@ -14,7 +14,7 @@ defmodule ExPlasma.Block do
   defstruct(hash: nil, transactions: [])
 
   alias ExPlasma.Encoding
-  alias ExPlasma.Transaction2, as: Transaction
+  alias ExPlasma.Transaction, as: Transaction
 
   # TODO
   #
@@ -24,13 +24,13 @@ defmodule ExPlasma.Block do
 
   ## Example
 
-  iex> %ExPlasma.Transaction2{tx_type: 1} |> List.wrap() |> ExPlasma.Block.new()
+  iex> %ExPlasma.Transaction{tx_type: 1} |> List.wrap() |> ExPlasma.Block.new()
   %ExPlasma.Block{
     hash: <<168, 54, 172, 201, 1, 212, 18, 167, 34, 57, 232, 89, 151, 225, 172,
       150, 208, 77, 194, 12, 174, 250, 146, 254, 93, 42, 28, 253, 203, 237, 247,
       62>>,
     transactions: [
-      %ExPlasma.Transaction2{
+      %ExPlasma.Transaction{
         inputs: [],
         metadata: <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
           0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>,
