+++ lib/ex_plasma/transaction/type/payment_v1.ex
@@ -1,9 +1,9 @@
-defmodule ExPlasma.Transaction2.Type.PaymentV1 do
+defmodule ExPlasma.Transaction.Type.PaymentV1 do
   @moduledoc false
 
-  @behaviour ExPlasma.Transaction2
+  @behaviour ExPlasma.Transaction
 
-  alias ExPlasma.Transaction2
+  alias ExPlasma.Transaction
 
   @type validation_responses() ::
           ExPlasma.Output.Type.PaymentV1.validation_responses()
@@ -26,11 +26,11 @@ defmodule ExPlasma.Transaction2.Type.PaymentV1 do
   ## Example
 
   iex> txn = %{inputs: [%{output_data: nil, output_id: %{blknum: 0, oindex: 0, position: 0, txindex: 0}, output_type: nil}], metadata: <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>, outputs: [%{output_data: %{amount: 1, output_guard: <<29, 246, 47, 41, 27, 46, 150, 159, 176, 132, 157, 153, 217, 206, 65, 226, 241, 55, 0, 110>>, token: <<46, 38, 45, 41, 28, 46, 150, 159, 176, 132, 157, 153, 217, 206, 65, 226, 241, 55, 0, 110>>}, output_id: nil, output_type: 1}], sigs: [], tx_data: <<0>>, tx_type: 1}
-  iex> ExPlasma.Transaction2.Type.PaymentV1.to_rlp(txn)
+  iex> ExPlasma.Transaction.Type.PaymentV1.to_rlp(txn)
   [[], <<1>>, [<<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>], [[<<1>>, [<<29, 246, 47, 41, 27, 46, 150, 159, 176, 132, 157, 153, 217, 206, 65, 226, 241, 55, 0, 110>>, <<46, 38, 45, 41, 28, 46, 150, 159, 176, 132, 157, 153, 217, 206, 65, 226, 241, 55, 0, 110>>, <<1>>]]], 0, <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>]
   """
-  @impl Transaction2
-  @spec to_rlp(Transaction2.t()) :: list()
+  @impl Transaction
+  @spec to_rlp(Transaction.t()) :: list()
   def to_rlp(%{} = transaction) do
     [
       transaction.sigs,
@@ -60,7 +60,7 @@ defmodule ExPlasma.Transaction2.Type.PaymentV1 do
   ...>  0,
   ...>  <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>
   ...>]
-  iex> ExPlasma.Transaction2.Type.PaymentV1.to_map(rlp)
+  iex> ExPlasma.Transaction.Type.PaymentV1.to_map(rlp)
   %{
   	inputs: [
   		%ExPlasma.Output{
@@ -88,8 +88,8 @@ defmodule ExPlasma.Transaction2.Type.PaymentV1 do
   	tx_type: 1
   }
   """
-  @impl Transaction2
-  @spec to_map(list()) :: Transaction2.t()
+  @impl Transaction
+  @spec to_map(list()) :: Transaction.t()
   def to_map(rlp), do: do_to_map(rlp)
 
   defp do_to_map([sigs, tx_type, inputs, outputs, "", metadata]),
@@ -115,9 +115,9 @@ defmodule ExPlasma.Transaction2.Type.PaymentV1 do
   ## Example
 
   iex> txn = %{inputs: [%{output_data: [], output_id: %{blknum: 0, oindex: 0, position: 0, txindex: 0}, output_type: nil}], metadata: <<0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0>>, outputs: [%{output_data: %{amount: <<0, 0, 0, 0, 0, 0, 0, 1>>, output_guard: <<29, 246, 47, 41, 27, 46, 150, 159, 176, 132, 157, 153, 217, 206, 65, 226, 241, 55, 0, 110>>, token: <<46, 38, 45, 41, 28, 46, 150, 159, 176, 132, 157, 153, 217, 206, 65, 226, 241, 55, 0, 110>>}, output_id: nil, output_type: 1}], sigs: [], tx_data: <<0>>, tx_type: <<1>>}
-  iex> {:ok, ^txn} = ExPlasma.Transaction2.Type.PaymentV1.validate(txn)
+  iex> {:ok, ^txn} = ExPlasma.Transaction.Type.PaymentV1.validate(txn)
   """
-  @impl Transaction2
+  @impl Transaction
   @spec validate(map()) :: validation_responses()
   def validate(%{} = transaction) do
     with {:ok, _inputs} <- do_validate_total(:inputs, transaction.inputs, 0),
